Fix / improve generate and --watch output and add test

+++ cli/prisma2/fixtures/test.sh
@@ -7,6 +7,15 @@ if [[ ${VERSION} != *"prisma2@"* ]]; then
   exit 1
 fi
 
+# Test generate output command
+cd fixtures/project/subdir
+GENERATE=$(node ../../../build/index.js generate)
+if [[ ${GENERATE} != *"Generated "* ]]; then
+  echo "prisma2 generate is broken"
+  exit 1
+fi
+cd ../../..
+
 
 # Test generation in npm script
 rm -rf fixtures/project/subdir/@prisma
